+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/router/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './router-container'
+export { default } from './container'
