+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/router/presentation.tsx
@@ -13,34 +13,32 @@
  *
  **/
 import * as React from 'react'
-import Router from '../../presentation/router'
+import styled from 'styled-components'
+import { CustomElement } from '../styles/mixins'
+import Routes from '../container/routes-container'
 
-import Navigation from '../../navigation'
-import ExtensionPoints from '../../../extension-points'
+const Router = styled.div`
+  ${CustomElement} overflow: hidden;
 
-type Props = {
-  navigation: React.ReactNode
-  routeDefinitions: object
+  > *:first-child {
+    height: calc(2 * ${props => props.theme.minimumLineSize});
+  }
+
+  > *:not(:first-child) {
+    height: calc(100% - 2 * ${props => props.theme.minimumLineSize});
+  }
+`
+
+interface Props {
+  nav: React.ReactNode
+  routeDefinitions: any
 }
 
-const Providers = ExtensionPoints.providers
-
-class RouterContainer extends React.Component<Props, {}> {
-  constructor(props: Props) {
-    super(props)
-  }
-  render() {
-    const navigation = <Navigation {...this.props} />
-    return (
-      <Providers>
-        <Router
-          nav={navigation}
-          routeDefinitions={this.props.routeDefinitions}
-          {...this.props}
-        />
-      </Providers>
-    )
-  }
+export default function(props: Props) {
+  return (
+    <Router>
+      {props.nav}
+      <Routes isMenu={false} routeDefinitions={props.routeDefinitions} />
+    </Router>
+  )
 }
-
-export default RouterContainer
